+++ derive/src/context/num/traits/identities.rs
@@ -4,7 +4,11 @@ use quote::quote;
 use syn::parse_quote;
 
 impl Context {
-    pub(crate) fn derive_zero(&self) -> proc_macro::TokenStream {
+    pub(crate) fn derive_zero(&self) -> proc_macro2::TokenStream {
+        if self.non_static_modulus {
+            return quote!();
+        }
+
         let Context {
             modulus,
             implementation,
@@ -21,7 +25,7 @@ impl Context {
         let zero = parse_quote!(<#implementation as #modtype::Impl>::zero(#modulus));
         let zero = self.struct_expr(true, Some(zero));
 
-        quote!(
+        quote! {
             impl#impl_generics #num_traits::Zero for #struct_ident#ty_generics
             #where_clause
             {
@@ -35,11 +39,14 @@ impl Context {
                     <#implementation as #modtype::Impl>::is_zero(self.#field_ident, #modulus)
                 }
             }
-        )
-        .into()
+        }
     }
 
-    pub(crate) fn derive_one(&self) -> proc_macro::TokenStream {
+    pub(crate) fn derive_one(&self) -> proc_macro2::TokenStream {
+        if self.non_static_modulus {
+            return quote!();
+        }
+
         let Context {
             modulus,
             implementation,
@@ -56,7 +63,7 @@ impl Context {
         let one = parse_quote!(<#implementation as #modtype::Impl>::one(#modulus));
         let one = self.struct_expr(true, Some(one));
 
-        quote!(
+        quote! {
             impl#impl_generics #num_traits::One for #struct_ident#ty_generics
             #where_clause
             {
@@ -70,7 +77,6 @@ impl Context {
                     <#implementation as #modtype::Impl>::is_one(self.#field_ident, #modulus)
                 }
             }
-        )
-        .into()
+        }
     }
 }