+++ derive/src/context/std/fmt.rs
@@ -4,7 +4,7 @@ use quote::quote;
 use syn::LitStr;
 
 impl Context {
-    pub(crate) fn derive_display(&self) -> proc_macro::TokenStream {
+    pub(crate) fn derive_display(&self) -> proc_macro2::TokenStream {
         let Context {
             modulus,
             implementation,
@@ -17,7 +17,7 @@ impl Context {
         } = self;
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
-        quote!(
+        quote! {
             impl#impl_generics #std::fmt::Display for #struct_ident#ty_generics
             #where_clause
             {
@@ -30,11 +30,10 @@ impl Context {
                     )
                 }
             }
-        )
-        .into()
+        }
     }
 
-    pub(crate) fn derive_debug(&self) -> proc_macro::TokenStream {
+    pub(crate) fn derive_debug(&self) -> proc_macro2::TokenStream {
         let Context {
             modulus,
             implementation,
@@ -49,7 +48,7 @@ impl Context {
 
         let tuple_name = LitStr::new(&struct_ident.to_string(), struct_ident.span());
 
-        quote!(
+        quote! {
             impl#impl_generics #std::fmt::Debug for #struct_ident#ty_generics
             #where_clause
             {
@@ -63,7 +62,6 @@ impl Context {
                     )
                 }
             }
-        )
-        .into()
+        }
     }
 }