+++ derive/src/context/std/marker.rs
@@ -3,7 +3,7 @@ use crate::context::Context;
 use quote::quote;
 
 impl Context {
-    pub(crate) fn derive_copy(&self) -> proc_macro::TokenStream {
+    pub(crate) fn derive_copy(&self) -> proc_macro2::TokenStream {
         let Self {
             std,
             struct_ident,
@@ -12,13 +12,12 @@ impl Context {
         } = self;
         let (impl_generics, ty_generics, where_clause) = generics.split_for_impl();
 
-        quote!(
+        quote! {
             #[automatically_derived]
             impl#impl_generics #std::marker::Copy for #struct_ident#ty_generics
             #where_clause
             {
             }
-        )
-        .into()
+        }
     }
 }