+++ derive/src/context/std/str.rs
@@ -3,7 +3,11 @@ use crate::context::Context;
 use quote::quote;
 
 impl Context {
-    pub(crate) fn derive_from_str(&self) -> proc_macro::TokenStream {
+    pub(crate) fn derive_from_str(&self) -> proc_macro2::TokenStream {
+        if self.non_static_modulus {
+            return quote!();
+        }
+
         let Context {
             modulus,
             implementation,
@@ -18,7 +22,7 @@ impl Context {
 
         let struct_expr = self.struct_expr(true, None);
 
-        quote!(
+        quote! {
             impl#impl_generics #std::str::FromStr for #struct_ident#ty_generics
             #where_clause
             {
@@ -30,7 +34,6 @@ impl Context {
                     #std::result::Result::Ok(#struct_expr)
                 }
             }
-        )
-        .into()
+        }
     }
 }